ecomate-ai/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.6
httpx==0.27.2
selectolax==0.3.20
beautifulsoup4==4.12.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
pint==0.24.4
python-slugify==8.0.4
boto3==1.34.152
psycopg[binary]==3.2.1
pgvector==0.3.3
temporalio==1.7.0
PyYAML==6.0.2
pdfplumber==0.11.4